Shellescape new annotations
Shellescape the annotation input to prevent errors when entering a
string with shell glob characters, such as question marks (?) or
asterisks (*).

The error can be observed by attempting to insert an annotation such as
"hello?world". Without the fix, the annotation simply fails to be added
to the task. You can make it more obvious as to what the underlying
issue is by changing the last argument of `system_call` from 'silent' to
'echo'.